Add files via upload
Version 2. Fixed: Moves are not hardcoded in the code. Updated table creation methods. Only the initial table is written in code, but it is increased using methods. Rock_Paper_Scissors, with the ability to select from 3 to 7 parameters directly in the program. There is a check for input and exit from the program.  Link to video on google drive: https://drive.google.com/drive/folders/1wsCH_t2Gb79UmEmWmP0zUIGUjDhKhnQU?usp=sharing